Change decision method for absolute vs relative.
Before, the condition for using an absolute jump was if 'number' is on.
In Vim 7.3 with patches 1-46, if 'number' is on, the line numbering must
be relative. In newer versions, 'number' can be on, but if
'relativenumber' is also on, the number at the current line is absolute
instead of zero.