Change decision method for absolute vs relative.
Before, the condition for using an absolute jump was if 'number' is on.
In Vim 7.3 with patches 1-46, if 'number' is on, the line numbering must
be absolute. In newer versions, 'number' being on is not an
unconditional indicator that the line numbering is absolute instead of
relative. If 'relativenumber' is also on, the line numbering is
relative: the number at the current line is absolute instead of zero.