Switch 1 and 2 for GetUserInput
GetUserInput() returns a list describing the user's input. The first
item is 0 for a cancel keystroke, or 1 or 2 depending on whether the
user wants to type a one- or two-digit number. Previously, this was 1
for a two-digit number, and 2 for a one-digit number. This seemed
illogical, so I switched it.